# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nex Club API
  version: 0.1.0
  summary: 内容服务公开接口与管理接口。未实现的路径返回 501，code 为 not_implemented。
servers:
  - url: /
paths:
...
  /api/admin/automation:
    get:
      operationId: automationOverview
      tags: [admin, automation]
      summary: 运行概览与公开的模型配置
      security: [{adminSession: []}]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/sources:
    get:
      operationId: listSources
      tags: [admin, automation]
      summary: 分页列出信源（不返回凭据）
      security: [{adminSession: []}]
      parameters:
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationPage"}
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createSource
      tags: [admin, automation]
      summary: 创建信源，不立即采集
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationSourceInput"}
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/sources/{id}:
    put:
      operationId: updateSource
      tags: [admin, automation]
      summary: 按 edit_version 更新信源；变更采集地址会清空旧检查点
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationSourceInput"}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/sources/{id}/run:
    post:
      operationId: queueSource
      tags: [admin, automation]
      summary: 立即采集；已有活跃采集任务时返回 409
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationVersionInput"}
      responses:
        "202":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/ingest-runs:
    get:
      operationId: listIngestRuns
      tags: [admin, automation]
      summary: 分页查询采集任务，支持 status 和 source_id
      security: [{adminSession: []}]
      parameters:
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: status
          in: query
          schema: {type: string}
        - name: source_id
          in: query
          schema: {type: string}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationPage"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/processing-runs:
    get:
      operationId: listProcessingRuns
      tags: [admin, automation]
      summary: 分页查询加工阶段，支持 status 和 source_id
      security: [{adminSession: []}]
      parameters:
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: status
          in: query
          schema: {type: string}
        - name: source_id
          in: query
          schema: {type: string}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationPage"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/processing-runs/{id}:
    get:
      operationId: processingContext
      tags: [admin, automation]
      summary: 原始资料与同轮次阶段；正文最多返回 30000 字符
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/processing-runs/{id}/retry:
    post:
      operationId: retryProcessing
      tags: [admin, automation]
      summary: 沿用冻结计划重试；未确认调用或活跃任务会阻止重试
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationEmpty"}
      responses:
        "202":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/processing-runs/{id}/rerun:
    post:
      operationId: rerunProcessing
      tags: [admin, automation]
      summary: 显式新建加工轮次，使用当前模型配置；可能产生费用
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationEmpty"}
      responses:
        "202":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/proposals:
    get:
      operationId: listAutomationProposals
      tags: [admin, automation]
      summary: 分页查询审核建议，支持 status 与 kind
      security: [{adminSession: []}]
      parameters:
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: status
          in: query
          schema: {type: string}
        - name: kind
          in: query
          schema: {type: string}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationPage"}
        default:
          $ref: "#/components/responses/Error"
  /api/admin/automation/calls:
    get:
      operationId: listAutomationCalls
      tags: [admin, automation]
      summary: 分页查询模型调用与费用，不返回密钥和提示词
      security: [{adminSession: []}]
      parameters:
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: status
          in: query
          schema: {type: string}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationPage"}
        default:
          $ref: "#/components/responses/Error"
...
  /api/admin/proposals/{id}/preview:
    post:
      operationId: previewProposal
      tags: [admin, automation]
      summary: 按审核决定生成预览，不保存、不发布、不记录访客事件
      security: [{adminSession: []}]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AutomationDecision"}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AutomationObject"}
        default:
          $ref: "#/components/responses/Error"
...
  /api/admin/source-capabilities:
    get:
      operationId: sourceCapabilities
      tags: [admin, automation]
      security: [{adminSession: []}]
      responses:
        "200":
          description: 付费渠道的凭据与预算是否已配置，仅返回状态和环境变量名。
        default: {$ref: "#/components/responses/Error"}
  /api/admin/source-presets:
    get:
      operationId: sourcePresets
      tags: [admin, automation]
      security: [{adminSession: []}]
      responses:
        "200":
          description: 固定版本的 AIHOT 示范清单、来源链接与已导入标记。
        default: {$ref: "#/components/responses/Error"}
  /api/admin/source-presets/import:
    post:
      operationId: importSourcePresets
      tags: [admin, automation]
      security: [{adminSession: []}]
      parameters:
        - {$ref: "#/components/parameters/csrf"}
        - {$ref: "#/components/parameters/idempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: {type: array, minItems: 1, maxItems: 18, items: {type: string}}
      responses:
        "200":
          description: created / existing 数量及新 source_ids。新信源均暂停，已有配置不覆盖。
        default: {$ref: "#/components/responses/Error"}
  /api/admin/sources/preview:
    post:
      operationId: previewSource
      tags: [admin, automation]
      security: [{adminSession: []}]
      description: 使用未保存的配置试抓，不写入资料、采集任务或加工任务。付费渠道保留调用回执和预算记录。
      parameters:
        - {$ref: "#/components/parameters/csrf"}
        - name: Idempotency-Key
          in: header
          description: 付费渠道必填；相同请求重放复用回执，不重新收费。
          schema: {type: string, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: {type: string, enum: [github, rss, json, web, x, wechat]}
                name: {type: string}
                repository: {type: string}
                feed_url: {type: string}
                config: {type: object, additionalProperties: true}
      responses:
        "200":
          description: total、最多 20 条 items、warnings、elapsed_ms；persisted_items=0、processing_started=false。
        "429":
          description: 同时试抓的请求过多。
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    mcpBearer:
      type: http
      scheme: bearer
    adminSession:
      type: apiKey
      in: cookie
      name: nex_session
  parameters:
...
    resourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: 错误信封。field_errors 没有时为 []。
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    AutomationObject:
      type: object
      additionalProperties: true
    AutomationPage:
      type: object
      required: [items, has_more, next_offset]
      properties:
        items: {type: array, maxItems: 50, items: {$ref: "#/components/schemas/AutomationObject"}}
        has_more: {type: boolean}
        next_offset: {type: integer, minimum: 0}
    AutomationEmpty:
      type: object
      additionalProperties: false
    AutomationVersionInput:
      type: object
      required: [edit_version]
      properties:
        edit_version: {type: integer, minimum: 1}
    AutomationSourceInput:
      type: object
      required: [name, kind, participation_mode, trust_tier, enabled, interval_seconds]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        kind: {type: string, enum: [github, rss, json, web, x, wechat, external]}
        config: {type: object, description: 经类型校验的公开采集配置，不接受密钥, additionalProperties: true}
        repository: {type: string, description: GitHub 信源必填 owner/name}
        feed_url: {type: string, description: RSS 信源必填公开 HTTP(S) 地址}
        participation_mode: {type: string, enum: [content, signal, internal]}
        trust_tier: {type: string, enum: [official, verified, community, excluded]}
        enabled: {type: boolean}
        interval_seconds: {type: integer, minimum: 300, maximum: 604800}
        allow_fulltext: {type: boolean, default: false}
        edit_version: {type: integer, description: 更新信源时必填当前版本}
    AutomationDecision:
      type: object
      required: [edit_version, fields]
      properties:
        edit_version: {type: integer, minimum: 0}
        fields:
          type: object
          description: 未提供的字段一律拒绝；受保护字段须显式解锁。
          additionalProperties: {type: string, enum: [accept, reject, rewrite]}
        rewrites: {type: object, additionalProperties: true}
        unlock: {type: array, items: {type: string}}
        reason: {type: string}
...
    Error:
      type: object
      required: [code, message, request_id, field_errors]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        field_errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
...